guedes/requirements.txt
appdirs==1.4.3
boto3==1.9.71
botocore==1.12.71
Django==1.11.17
django-appconf==1.0.2
django-bakery==0.12.7
django-ckeditor==5.6.1
django-classy-tags==0.8.0
django-colorful==1.2
django-compressor==2.1.1
django-debug-toolbar==1.6
django-extensions==1.7.6
django-filer==1.2.6
django-flat-theme==1.1.4
django-js-asset==1.1.0
django-maintenancemode==0.11.2
django-memcached==0.1.2
django-mptt==0.8.7
django-polymorphic==1.0.2
django-robots==2.0
django-sekizai==0.10.0
django-solo==1.1.2
docutils==0.14
easy-thumbnails==2.5
fs==2.1.3
IPy==0.83
jmespath==0.9.3
mysqlclient==1.3.9
olefile==0.44
packaging==16.8
Pillow==5.3.0
pkg-resources==0.0.0
pyparsing==2.1.10
python-dateutil==2.7.5
python-memcached==1.58
pytz==2018.7
rcssmin==1.0.6
rjsmin==1.0.12
s3transfer==0.1.13
six==1.10.0
sqlparse==0.2.2
typing==3.6.6
Unidecode==0.4.20
urllib3==1.24.1
